Mark selected items as done
For the user is it convenient, when he can mark items as done, with
an simple click. Therefore implement this behavior so. But for that
reason, we need another way to edit the item, to solve this issue,
on an easy way, just add an edit button.

Additionally: change the layout, so that we have an simple grid
layout on the left side the item, and on the right side the edit
button. So that the user-interface ist neat.