Restart the app development
In the meantime angular gets a lot more features. For example the update
from 2.4 to 4.2.4, and the angular-cli instead of self written grunt tasks.
The main-reason for this grocery app is to lern technices, and to play around.
Therefore make an clean cut and rewrite this app.

- Delete the old angular app.
- Remain an few files, like the license or cordova stuff, because they are
  still the same.
- Generate with the cli an new base structure (currently empty).
-> It's now an empty app withouth any functionality.

The next steps would be to reimplement all features, so that's hopefully it
works at the end like before.